// client/client/src/components/favorite-list/favorite.dto.ts
export interface IFavorite {
	BOOK_ID?: string;
	TITLE?: string;
	DESCRIPTION?: string;
	PUBLICATION_DATE?: Date;
	PUBLISHER_ID?: string;
	ISBN?: string;
	TAGS?: string;
	NUM_PAGES?: string;
	COPIES?: string;
	IMAGE_NAME?: string;
	FILE_NAME?: string;
	AUTHOR_NAME?: string;
	CATEGORY_NAME?: string;
}